+++ src/constructs/__tests__/character-class.test.ts
