+++ src/constructs/character-class.ts
